Consider the current stage when applying Action::Complete to the vqueue meta
When applying Action::Complete to the vqueue meta we need to consider what's the current stage
to properly update the statistics and the metadata. Hence, we are now passing the stage instead of whether it was in the inbox or not because we need to handle `Stage::Run` and `Stage::Park` differently.