Set InvocationStatus correctly when resuming invocation with vqueues
When resuming an invocation using the vqueue scheduler we cannot go directly into
the InvocationStatus::Invoked since we first need to get scheduled again. As part
of the scheduling the system needs to go through a few state changes that depend
on the InvocationStatus. Hence, this commit sets the InvocationStatus to Inboxed
when moving an item back into the inbox stage.