Temporarily avoid InvocationStatus::Invoked to ::Inboxed state transition when using vqueues
While it is correct that we should go from InvocationStatus::Invoked to ::Inboxed when resuming
an invocation using vqueues, it is currently supported by the state machine. For it to be fully
supported, we need to support handling journal completions when being ::Inboxed. Morever, we
need to handle cancellations of ::Inboxed items depending on whether the invocation was running
before or not. Additionally, we need to support the hotfix_apply_cancellation_after_deployment_is_pinned
when going from ::Suspended to ::Inbox, for example. Until this is done, we stay in ::Invoked when
resuming an invocation even when using vqueues.